% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenfuns.R
\name{eigenfuns}
\alias{eigenfuns}
\alias{eigenfuns.ate}
\alias{eigenfuns.pctn}
\alias{eigenfuns.tef}
\title{Extract temporal eigenfunctions}
\usage{
eigenfuns(x, ...)

\method{eigenfuns}{ate}(x, take = NULL, ...)

\method{eigenfuns}{pctn}(x, take = NULL, ...)

\method{eigenfuns}{tef}(x, take = NULL, ...)
}
\arguments{
\item{x}{the object from which to extract temporal eigenfunctions}

\item{...}{additional arguments passed to methods}

\item{take}{numeric; which eigenfunctions should be returned}
}
\value{
A data frame with one column per eigenfunction
}
\description{
Simple extractor function to access temporal eigenfunctions from various objects. Currently methods for class(es) `"ate"`, `"pctn"`, and `"tef"` are provided.
}
\author{
Gavin L. Simpson
}
