% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeWeightMat}
\alias{makeWeightMat}
\title{Diagonal weight matrix for time points}
\usage{
makeWeightMat(tp, FUN = NULL, link0 = TRUE, dimnames = TRUE, ...)
}
\arguments{
\item{tp}{numeric vector of sorted time points}

\item{FUN}{a function to apply}

\item{link0}{logical; should the link from t[0] be included?}

\item{dimnames}{logical; should dimnames by attached to the
matrix?}

\item{...}{optional arguments passed to \code{FUN}}
}
\value{
a diagonal matrix of weights with dimensions \code{length(tp) - 1}
}
\description{
A diagonal matrix of weights based on the inverse
of time duration between time points.
}
\examples{
tp <- seq_len(10)
makeWeightMat(tp)
}
\author{
Gavin L. Simpson
}
\keyword{utilities}
