% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{plot.ate}
\alias{plot.ate}
\title{Plot asymmetric temporal eigenfunctions}
\usage{
\method{plot}{ate}(x, pages = 1, ylim, ylab = nams, xlab = "",
  ask = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"ate"}}

\item{pages}{numeric; the number of pages over which to spread
the plots of the individual eigenfunction}

\item{ylim}{numeric vector of limits for the y-axis}

\item{xlab, ylab}{x and y-axis labels}

\item{ask}{logical; should plotting be paused between pages?}

\item{...}{additional arguments passed to \code{\link{plot}}}
}
\value{
A plot on the currently active device
}
\examples{
tp <- seq_len(50)
mod <- ate(tp)
plot(mod, pages = 2)

}
\seealso{
\code{\link{ate}} for creating ATE objects
}
\author{
Gavin L. Simpson
}
\keyword{hplot}
