% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelag.R
\name{plot.timelagAnalysis}
\alias{plot.timelagAnalysis}
\title{A time lag analysis plot.}
\usage{
\method{plot}{timelagAnalysis}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"timelagAnalysis"}, the result of
a call to \code{\link{timelag}}.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
Invisibly returns its input \code{x}.
}
\description{
A plot of \eqn{\sqrt{\mathrm{lag}}}{sqrt(lag)} against
compositional dissimilarity.
}
\details{
TODO - smoothers etc.
}
\examples{
## load analogue for Abernethy data set & distance()
if (require("analogue")) {

## Load Abernethy Forest data set
data("abernethy", package = "analogue")
## Load Abernethy Forest data set

## Remove the Depth and Age variables
abernethy2 <- abernethy[, -(37:38)]

## time lag analysis
dij <- as.dist(distance(abernethy2, method = "chord"))
tla <- timelag(dij)
plot(tla, pch = 19)
}
}
\author{
Gavin L. Simpson
}
