\name{moranI}
\alias{moranI}
\alias{moranI.ate}
\title{Moran's I for temporal eigenfunctions}
\usage{
moranI(x, ...)

\method{moranI}{ate}(x, ...)
}
\arguments{
  \item{x}{an R object of class \code{"ate"}.}

  \item{...}{additional arguments passed to methods.}
}
\value{
Numeric vector of Moran's I statistics
}
\description{
Moran's I for temporal eigenfunctions
}
\author{
Gavin L. Simpson
}

