\name{plot.moranI}
\alias{plot.moranI}
\title{Plot Moran's I statistics for temporal eigenfunctions}
\usage{
\method{plot}{moranI}(x, type = "b", xlab = "Eigenfunctions",
  ylab = "Moran's I", ...)
}
\arguments{
  \item{x}{an object of class \code{\link{moranI}}}

  \item{type}{}

  \item{...}{}
}
\value{
A plot on the current device
}
\description{
A plot of Moran's I versus time
}
\author{
Gavin
}

